add-viewlib: rename settings-related modules for consistency
