+++ _mur/web-client/@web-index.ts
@@ -14,7 +14,7 @@ export * as ViewLib from './viewlib/index.ts';
 /// interop with NetCreate
 export * as NCI from './nc-client-interop.ts';
 /// settings client
-export * as Settings from './settings-client.ts';
+export * as Settings from './mur-settings-client.ts';
 
 /// FORWARDED COMMON EXPORTS //////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
