+++ app/view/netcreate/components/react-settings-client.js
